jira-fixes.user.js: Disconnect MutationObserver
We only need the observer once to focus the left column on page load.
After that, we don't need to be listening to DOM changes.

Without disconnecting the observer, this broke '@'-name completion to
reference other users. Because that completion adds a new element (the
user list picker) to the column, our mutation observer gets called, and
we focus the left column, un-focussing the comment area.

Glad to have finally discovered why '@'-name completion in Jira comments
broke for me.